+++ src/main/java/io/xpring/xrpl/XRPClientDecorator.java
@@ -53,7 +53,7 @@ public interface XRPClientDecorator {
      * Retrieve the raw transaction status for the given transaction hash.
      *
      * @param transactionHash: The hash of the transaction.
-     * @return an {@link io.xpring.proto.TransactionStatus} containing the raw transaction status.
+     * @return an {@link RawTransactionStatus} containing the raw transaction status.
      */
     RawTransactionStatus getRawTransactionStatus(String transactionHash) throws XRPException;